backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5
psycopg2-binary==2.9.9
authlib==1.3.0
requests==2.31.0
PyGithub==2.1.1
atlassian-python-api==3.41.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
google-generativeai==0.3.1
pyyaml==6.0.1
docker==7.0.0
cryptography==41.0.7
gunicorn==21.2.0
python-socketio==5.10.0
eventlet==0.33.3
paramiko==3.4.0
